+++ rubato/classes/components/__init__.py
@@ -6,6 +6,7 @@ from .hitbox import Hitbox, Polygon, Circle, Rectangle
 from .rigidbody import RigidBody
 from .text import Text
 from .button import Button
+from .raster import Raster
 from .image import Image
 from .animation import Animation
 from .spritesheet import Spritesheet